put functions into files. Not sure if i did it well
