+++ car_rental_login_system/car_rental_login_system.cpp
@@ -1,294 +1,12 @@
-#include <iostream>
-#include <string>
-#include <vector>
+#include "access.h"
+#include "cars.h"
+#include <limits>
 #include <fstream>
-#include<ios>
+#include <ios>
+#include <cctype>
 #include <algorithm>
 
-class Ferrari_Car
-{
-    public:
-        std::string Name;
-        std::string engine;
-        int year;
-        int horespower;
-        bool four_wheel_drive;
-        int top_speed;
-        //int base_price; maybe
-        Ferrari_Car();
-        Ferrari_Car(std::string Name, std::string engine, int year, int horespower, bool four_wheel_drive, int top_speed)
-        {
-            this->Name = Name;
-            this->engine = engine;
-            this->year = year;
-            this->horespower = horespower;
-            this->four_wheel_drive = four_wheel_drive;
-            this->top_speed = top_speed;
-        }
-};
 //bunch of avaliable ferrari cars
-Ferrari_Car Superfast{ "Ferrari 812 Superfast", "6.5L V12", 2017, 789, false, 340};
-Ferrari_Car GTS{ "Ferrari 812 GTS", "6.2L nat asp V12", 2019, 789, false, 338};
-Ferrari_Car Stradale{ "Ferrari SF90 Stradale", "4.0L TT V8 + 3 electric", 2019, 989, true, 340};
-Ferrari_Car SFSpider{ "Ferrari SF90 Spider", "4.0L TT V8 + 3 electric", 2020, 989, false, 340};
-Ferrari_Car Tributo{ "Ferrari F8 Tributo", "3.9L twin-turbo V8", 2019, 710, false, 340};
-Ferrari_Car F8Spider{ "Ferrari F8 Spider", "3.9L twin-turbo V8", 2019, 710, false, 340};
-Ferrari_Car Roma{ "Ferrari Roma", "3.9L twin-turbo V8", 2020, 612, false, 320};
-Ferrari_Car Portofino{ "Ferrari Portofino M", "3.9L twin-turbo V8", 2017, 612, false, 320};
-Ferrari_Car MonzaSP1{ "Ferrari Monza SP1", "6.5 L nat asp V12", 2019, 810, false, 299};
-Ferrari_Car MonzaSP2{ "Ferrari Monza SP2", "6.5 L nat asp V12" , 2019, 810, false, 299};
-
-std::ostream& operator << (std::ostream& output, Ferrari_Car car) //allows us to print the car stats
-{
-    output << "\nName: " << car.Name << "\nEngine type: " << car.engine << "\nYear made: " << car.year << "\nHorsepower: " << car.horespower << "\nFour Wheel Drive: " << std::boolalpha << car.four_wheel_drive << "\nTop Speed: " << car.top_speed << "km/h\n";
-    return output;
-}
-
-void rent_car(std::vector<std::string> availablecars, std::string owner, std::vector<std::string> carlist)
-{
-   bool exit = false; //allows the loop below to be exited, only occurs if car is rented or no cars are available
-   if (availablecars.size() == 0)
-   {
-       exit = true;
-   }
-   else
-   {
-        std::cout << "Choose an car to see its stats and if the car is available, you can rent it: \n";
-   }
-    while(exit == false) //rents the given car
-    {
-        std::string car_to_view = "";
-        std::getline(std::cin, car_to_view); //gets the car that the user wants to hire
-        std::string car_name;
-        if (car_to_view == "Ferrari 812 Superfast" || car_to_view == "ferrari 812 superfast")
-        {
-            std::cout << Superfast;
-            car_name = "Ferrari 812 Superfast"; //sets the car name that the program needs to find
-        }
-        else if (car_to_view == "Ferrari 812 GTS" || car_to_view == "ferrari 812 gts" || car_to_view == "ferrari 812 GTS")
-        {
-            std::cout << GTS;
-            car_name = "Ferrari 812 GTS";
-        }
-        else if (car_to_view == "Ferrari SF90 Stradale" || car_to_view == "ferrari sf90 stradale" || car_to_view == "ferrari SF90 stradale")
-        {
-            std::cout << Stradale;
-            car_name = "Ferrari SF90 Stradale";
-        }
-        else if (car_to_view == "Ferrari SF90 Spider" || car_to_view == "ferrari sf90 spider" || car_to_view == "ferrari SF90 spider")
-        {
-            std::cout << SFSpider;
-            car_name = "Ferrari SF90 Spider";
-        }
-        else if (car_to_view == "Ferrari F8 Tributo" || car_to_view == "ferrari f8 tributo" || car_to_view == "ferrari F8 tributo")
-        {
-            std::cout << Tributo;
-            car_name = "Ferrari F8 Tributo";
-        }
-        else if (car_to_view == "Ferrari F8 Spider" || car_to_view == "ferrari f8 spider" || car_to_view == "ferrari F8 spider")
-        {
-            std::cout << F8Spider;
-            car_name = "Ferrari F8 Spider";
-        }
-        else if (car_to_view == "Ferrari Roma" || car_to_view == "ferrari roma")
-        {
-            std::cout << Roma;
-            car_name = "Ferrari Roma";
-        }
-        else if (car_to_view == "Ferrari Portofino M" || car_to_view == "ferrari portofino m" || car_to_view == "ferrari portofino M")
-        {
-            std::cout << Portofino;
-            car_name = "Ferrari Portofino M";
-        }
-        else if (car_to_view == "Ferrari Monza SP1" || car_to_view == "ferrari monza sp1" || car_to_view == "ferrari monza SP1")
-        {
-            std::cout << MonzaSP1;
-            car_name = "Ferrari Monza SP1";
-        }
-        else if (car_to_view == "Ferrari Monza SP2" || car_to_view == "ferrari monza sp2" || car_to_view == "ferrari monza SP2")
-        {
-            std::cout << MonzaSP2;
-            car_name = "Ferrari Monza SP2";
-        }
-        if (std::find(availablecars.begin(), availablecars.end(), car_name) != availablecars.end()) //if the car is available
-        {
-            std::cout << "Do you want to rent this car? Please type yes or no\n";
-            std::string rent_car;
-            std::getline(std::cin, rent_car); //gets the car that the user wants to rent
-            if (rent_car == "yes" || rent_car == "Yes")
-            {
-                std::fstream file(car_name, std::ios::app);
-                file << owner; //in the car file, says the car has been taken by this user
-                std::cout << "Well done, the car has been succesfully rented\n";
-                std::fstream file2(owner, std::ios::app);
-                file2 << "\n" << car_name; //puts the car name in the user file
-                exit = true; //allows the loop to be exited
-            }
-            else if (rent_car == "no" || rent_car == "No")
-            {
-                std::cout << "Okay, please type another car's name to view its stats\n"; //if user does not want to rent this car
-            }
-            else
-            {
-                std::cout << "Invalid answer, please type a car's name to view its stats\n";
-            }
-        }
-        else if(std::find(carlist.begin(), carlist.end(), car_name) != carlist.end())//if the car exists but is unavailable
-        {
-            std::cout << "This car is unavailable, so you can not rent it, please type another car to view its stats: \n";
-        }
-        else
-        {
-            std::cout << "That is not a valid car name, please type out the car name fully\n";
-        }
-    }
-
-}
-
-void returning_process(std::vector<std::string> user_details)
-{
-    std::string user_answer;
-    std::cin >> user_answer; //gets the user answer if they want to return their car
-
-    if (user_answer == "yes" || user_answer == "Yes")
-    {
-        std::ofstream file(user_details[0]);
-        file << user_details[0] << "\n" << user_details[1]; //recreates user file as empty
-        std::ofstream file1(user_details[2]); //recreate car file as empty
-        std::cout << "Thank you for returning your car\n"; //thank the user for logging in
-    }
-    else if (user_answer == "no" || user_answer == "No")
-    {
-        std::cout << "Okay, thank you for logging in\n";
-    }
-    else 
-    {
-        std::cout << "Invalid answer, redirecting back to starting page\n";
-    }
-}
-
-void registration_process(std::vector<std::string> unnacceptable_words)
-{
-    bool exit_loop = false;
-    std::string username;
-    std::string password;
-    do //is stuck in a loop until the inputs are acceptable
-    {
-        char variablecatcher = 0; //is used to catch an exception in the username
-        std::cin.ignore(std::numeric_limits<std::streamsize>::max(), '\n'); //clears cin cooldown, as the program was skipping over this line
-        std::cout << "\nUsername: ";
-        std::getline(std::cin, username); //gets username
-        std::cout << "\nPassword: ";
-        std::getline(std::cin, password); //gets password
-        for (char i = 0; i < unnacceptable_words.size(); i++)
-        {
-            if (username.find(unnacceptable_words[i]) != std::string::npos) //sees if there are unnacceptable sequences in username
-            {
-                std::cout << "the \"" << unnacceptable_words[i] << "\" is not allowed in username, please press enter to try again\n";
-                variablecatcher = 1;
-            }
-            if (password.find(unnacceptable_words[i]) != std::string::npos) //sees if there are unnacceptable sequences in username
-            {
-                std::cout << "the \"" << unnacceptable_words[i] << "\" is not allowed in password, please press enter to try again\n";
-                variablecatcher = 1;
-            }
-        }
-        variablecatcher == 0 ? exit_loop = true : exit_loop = false;
-    } while (exit_loop == false);
-    std::cout << "congrats, you have been registered\n";
-    std::ofstream file(username); //creates new file for every user that is registered
-    file << username << "\n" << password; //add username and password to file
-}
-
-void renting_process(std::string owner) //user can rent a car
-{
-    std::string user_answer;
-    std::cin >> user_answer;
-    std::vector<std::string> carlist = { "Ferrari 812 Superfast", "Ferrari 812 GTS", "Ferrari SF90 Stradale", "Ferrari SF90 Spider", "Ferrari F8 Tributo", "Ferrari F8 Spider", "Ferrari Roma", "Ferrari Portofino M", "Ferrari Monza SP1", "Ferrari Monza SP2" };
-    if (user_answer == "yes" || user_answer == "Yes") //if user answers yes
-    {
-        int current_car_from_carlist = 0;
-         std::vector<std::string> availablecars; //holds an array of the cars that are available to be rented
-        for (current_car_from_carlist; current_car_from_carlist < carlist.size(); ++current_car_from_carlist)
-        {
-            std::string available_car = "";
-            std::ifstream myfile(carlist[current_car_from_carlist], std::fstream::app); //creates file of the car if not created and opens it if it is
-            std::getline(myfile, available_car); //reads from file, if it has extra words then car is not available
-            if (available_car == "") //if there is nothing read
-            {
-                availablecars.push_back(carlist[current_car_from_carlist]);
-                std::cout << "The " << carlist[current_car_from_carlist] << " is available\n";
-            }
-            else
-            {
-                std::cout << "The " << carlist[current_car_from_carlist] << " is taken by a person with the username " << available_car << ".\n";
-            }
-        }
-        bool car_chosen = false;
-        std::cin.ignore(std::numeric_limits<std::streamsize>::max(), '\n'); //clears cin cooldown, as the program was skipping over this line
-        rent_car(availablecars, owner, carlist);
-        if (availablecars.size() == 0) //if the size is nothign
-        {
-            std::cout << "There are no cars available, please check again later\n";
-        }
-    }
-    else if (user_answer == "no" || user_answer == "No") //if user answers no
-    {
-        std::cout << "Okay, thank you for logging in\n";
-    }
-    else
-    {
-        std::cout << "Invalid answer, redirecting back to starting page\n";
-    }
-}
-
-void car_interaction(std::string owner) //allows user to buy or sell cars
-{
-    std::ifstream file(owner); //opens file of user
-    std::vector<std::string> user_details;
-    for (size_t i = 0; i < 9; i++) //gets the details on the note we have on the user
-    {
-        std::string read_from_file = "";
-        std::getline(file, read_from_file);
-        if (read_from_file != "") //if something is read, then push back
-        {
-            user_details.push_back(read_from_file);
-        }
-    }
-    if (user_details.size() == 2)
-    {
-        std::cout << "Do you want to rent a car? Please type Yes or No\n";
-        renting_process(owner);
-    }
-    else
-    {
-        std::cout << "Do you want to return your " << user_details[2] << "? Please type Yes or No\n";
-        returning_process(user_details);
-    }
-}
-
-void login_process()
-{
-    std::string username, password;
-    std::cin.ignore(std::numeric_limits<std::streamsize>::max(), '\n'); //clears cin cooldown, as the program was skipping over this line
-    std::cout << "\nUsername: ";
-    std::getline(std::cin, username); //gets username
-    std::cout << "\nPassword: ";
-    std::getline(std::cin, password); //gets password
-    std::ifstream file(username); //opens file with username login
-    std::string user, pass; //used to get data from the file
-    std::getline(file, user); //gets username on first line
-    std::getline(file, pass); //gets password on next line
-    if (user == username && pass == password) //mini hotfix, is meant to be && but username isn't being read
-    {
-        std::cout << "you have logged in, welcome " << username << "\n";
-        car_interaction(user); //allows user to buy or sell cars
-    }
-    if (user != username || pass != password)
-    {
-        std::cout << "invalid username or password\n";
-    }
-}
 
 int main()
 {
@@ -298,7 +16,7 @@ int main()
         std::string choice; //will recieve choice, would have made integer, but string provides for any input
         std::cout << "your choice: ";
         std::cin >> choice; //recieves choice
-        std::vector<std::string> unnacceptable_words = { " ", "_", "*", "&", "=", "'", "-", "\"", "-","+",",", "<", ">", "." };
+        std::vector<char> unnacceptable_words = { ' ', '_', '*', '&', '=', '\'', '-', '\'', '-','+',',', '<', '>', '.' };
         if (choice == "1")
         {
             login_process();
@@ -313,4 +31,4 @@ int main()
         }
 
     }
-}
+}
