+++ car_rental_login_system/string.h
@@ -1,0 +1,10 @@
+#pragma once
+#include "access.h"
+
+std::string stringLower(std::string &str) {
+    std::string lower("");
+    for (char &c : str) {
+        lower += c;
+    }
+    return lower;
+}
